Perform test locally instead of using a live dns resolution (#11)
Tests were relying on an external domain to be configured properly.
Especially, it needed to have both an A and AAAA record configured. It
turns out this domain has changed and no longer has an AAAA record,
making the tests to fail.

This commit updates the tests to use a local DoH resolver, which can be
configured to return deterministic answers. This way, the resolution
does not depend on an external configuration.